+++ CHANGELOG.md
@@ -38,6 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow option to add exercise at any position in training session
 - Keep search term on exercises and routines pages when going back in history
 - Display recent and previous exercises separately
+- Improve delete dialog by adding name or date
 
 ### Fixed